move validation function into routes/validation.ts
