+++ server/routes/accommodation/index.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express'
 import { Services } from '../../services'
 import AccommodationController from './accommodationController'
-import { getValidationForPathwayQuery } from '../../utils/utils'
+import { getValidationForPathwayQuery } from '../validation'
 
 export default (router: Router, services: Services) => {
   const accommodationController = new AccommodationController(services.rpService, services.prisonerDetailsService)
