+++ server/routes/attitudes-thinking-behaviour/index.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express'
 import { Services } from '../../services'
 import AttitudesThinkingBehaviourController from './attitudesThinkingBehaviourController'
-import { getValidationForPathwayQuery } from '../../utils/utils'
+import { getValidationForPathwayQuery } from '../validation'
 
 export default (router: Router, services: Services) => {
   const attitudesThinkingBehaviourController = new AttitudesThinkingBehaviourController(
