+++ server/routes/children-families-and-communities/index.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express'
 import { Services } from '../../services'
 import ChildrenFamiliesCommunitiesController from './childrenFamiliesCommunitiesController'
-import { getValidationForPathwayQuery } from '../../utils/utils'
+import { getValidationForPathwayQuery } from '../validation'
 
 export default (router: Router, services: Services) => {
   const childrenFamiliesCommunitiesController = new ChildrenFamiliesCommunitiesController(
