+++ server/routes/drugs-alcohol/index.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express'
 import { Services } from '../../services'
 import DrugsAlcoholController from './drugsAlcoholController'
-import { getValidationForPathwayQuery } from '../../utils/utils'
+import { getValidationForPathwayQuery } from '../validation'
 
 export default (router: Router, services: Services) => {
   const drugsAlcoholController = new DrugsAlcoholController(services.rpService, services.prisonerDetailsService)
