+++ server/routes/education-skills-work/index.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express'
 import { Services } from '../../services'
 import EducationSkillsWorkController from './educationSkillsWorkController'
-import { getValidationForPathwayQuery } from '../../utils/utils'
+import { getValidationForPathwayQuery } from '../validation'
 
 export default (router: Router, services: Services) => {
   const educationSkillsWorkController = new EducationSkillsWorkController(
