+++ server/routes/health-status/index.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express'
 import { Services } from '../../services'
 import HealthStatusController from './healthStatusController'
-import { getValidationForPathwayQuery } from '../../utils/utils'
+import { getValidationForPathwayQuery } from '../validation'
 
 export default (router: Router, services: Services) => {
   const healthController = new HealthStatusController(services.rpService, services.prisonerDetailsService)
